Make the code compile on a modern MacOS.
A few things had bitrotted since 2009:

* Fixed strndup() to have a size_t argument instead of int (compiler
  complained about a bad type).

* Fixed memmem() to cast memchr() result to (void*) (otherwise you get
  const void* and compiler does not like that).

* Added static_cast<int>s into parse_reqs(), as it would refuse to
  compile since it had implicit conversion from uint32_t to int.

* Added #include <unistd.h> to the input-mmap plugin as it did not know
  what close() was.

* Added hacks to Makefile to compile parser.cc and misc.c right into the
  plugins, because they complained about missing symbols.

I put "25May2025 Mikko Hack" and some explanations into parts that I
shamelessly did the minimum work required to get the project to compile.

I tested that it runs at all, but I can't say if you could actually load
plugins and have this work without issue. This commit is purely to make
typing "make" actually finish.

Did not test on Linux, just on my modern Sequoia 15.5 Mac. Clang was my
compiler.

The project is also super old so might not work anymore for its actual
intended purpose. I am interested in code archeology sense :)

+++ misc.h
@@ -54,7 +54,7 @@ void * memmem(const void *, size_t, const void *, size_t);
 #endif /* !HAVE_MEMMEM */
 
 #ifndef HAVE_STRNDUP
-char * strndup(const char *, unsigned int);
+char * strndup(const char *, size_t);
 #endif /* !HAVE_STRNDUP */
 
 #ifndef HAVE_ASPRINTF
